great

+++ routes/web.php
@@ -17,6 +17,8 @@ Route::get('/', function () {
 
 Route::get('/index', 'IndexController@index')->name('index');
 Route::get('/product/{id}', 'IndexController@productDetail')->name('productDetail');
+Route::get('/category', 'IndexController@getCategory')->name('getCategory');
+Route::get('/category/{id}', 'IndexController@getProductByCategory')->name('getProductByCategory');
 
 Route::get('/cart', 'CartController@getCart')->name('getCart');
 Route::post('/addCart/{id}', 'CartController@addCart')->name('addCart');
@@ -26,11 +28,20 @@ Route::get('/removeCart/{id}','CartController@removeCart')->where('id','[0-9]+')
 Route::get('/orderConfirm', 'CartController@orderConfirm')->name('orderConfirm');
 Route::post('/orderPay', 'CartController@orderPay')->name('orderPay');
 
-Route::prefix('admin')->middleware('admin')->group(function () {
+Route::get('/order_manager/','HomeController@order_manager')->name('order_manager');
+Route::get('/order_manager/details/{user_name}/{id}','HomeController@order_manager_detail')->name('order_manager_detail');
+
+Route::get('/profile/','HomeController@profile_manager')->name('profile_manager');
+Route::put('/profile/{id}','HomeController@update_profile_manager')->name('update_profile_manager');
+
+Route::get('/change-password/','HomeController@change_password')->name('change_password');
+Route::put('/change-password/{id}','HomeController@update_change_password')->name('update_change_password');
+
+Route::group(['prefix'=>'admin','middleware'=>'adminPage'],function () {
 
     Route::get('/dashboard', 'AdminController@index')->name('admin.index');
 
-    Route::prefix('categories')->group(function () {
+    Route::group(['prefix'=>'categories','middleware'=>'admin'],function () {
         Route::get('/list', 'CategoryController@index')->name('category.index');
         Route::get('/create', 'CategoryController@create')->name('category.create');
         Route::post('/create', 'CategoryController@store')->name('category.store');
@@ -40,7 +51,7 @@ Route::prefix('admin')->middleware('admin')->group(function () {
         Route::get('/search', 'CategoryController@search')->name('category.search');
     });
 
-    Route::prefix('products')->group(function () {
+    Route::group(['prefix'=>'products','middleware'=>'admin'],function () {
         Route::get('/list', 'ProductController@index')->name('product.index');
         Route::get('/show/{id}', 'ProductController@show')->name('product.show');
         Route::get('/create', 'ProductController@create')->name('product.create');
@@ -49,9 +60,14 @@ Route::prefix('admin')->middleware('admin')->group(function () {
         Route::put('/edit/{id}', 'ProductController@update')->name('product.update');
         Route::delete('/destroy/{id}', 'ProductController@destroy')->name('product.destroy');
         Route::get('/search', 'ProductController@search')->name('product.search');
+        Route::get('/images/{id}', 'ProductController@showImage')->name('product.getImage');
+        Route::delete('/destroy/image/{id}', 'ProductController@destroyImage')->name('product.destroyImage');
+        Route::get('/add/image/{id}', 'ProductController@createImage')->name('product.createImage');
+        Route::post('/add/image/{id}', 'ProductController@storeImage')->name('product.storeImage');
+        Route::put('/edit/image/{id}', 'ProductController@updateImage')->name('product.updateImage');
     });
 
-    Route::prefix('users')->group(function () {
+    Route::group(['prefix'=>'users','middleware'=>'admin'],function () {
         Route::get('/list', 'UserController@index')->name('user.index');
         Route::get('/show/{id}', 'UserController@show')->name('user.show');
         Route::get('/create', 'UserController@create')->name('user.create');
@@ -64,23 +80,23 @@ Route::prefix('admin')->middleware('admin')->group(function () {
         Route::post('/decentralization/{id}', 'UserController@decentralization')->name('user.decentralization');
     });
 
-    Route::prefix('orders')->group(function () {
-        Route::get('/list', 'OrderController@index')->name('order.index');
+    Route::group(['prefix'=>'orders'],function () {
+        Route::get('/list', 'OrderController@index')->middleware('admin')->name('order.index');
         Route::get('/show/{id}', 'OrderController@show')->name('order.show');
         Route::get('/search', 'OrderController@search')->name('order.search');
 
-        Route::delete('/cancel/{id}', 'OrderController@cancel')->name('order.cancel');
-        Route::get('/cancel', 'OrderController@getCancel')->name('order.getCancel');
-        Route::post('/cancel/{id}', 'OrderController@restoreOrders')->name('order.restore');
+        Route::delete('/cancel/{id}', 'OrderController@cancel')->middleware('admin')->name('order.cancel');
+        Route::get('/cancel', 'OrderController@getCancel')->middleware('admin')->name('order.getCancel');
+        Route::post('/cancel/{id}', 'OrderController@restoreOrders')->middleware('admin')->name('order.restore');
 
-        Route::get('/processing', 'OrderController@getProcessing')->name('order.getProcessing');
-        Route::put('/processing/{id}', 'OrderController@processing')->name('order.processing');
+        Route::get('/processing', 'OrderController@getProcessing')->middleware('admin')->name('order.getProcessing');
+        Route::put('/processing/{id}', 'OrderController@processing')->middleware('admin')->name('order.processing');
 
-        Route::get('/export', 'OrderController@getExportOrder')->name('order.getExport');
-        Route::put('/export/{id}', 'OrderController@exportOrder')->name('order.export');
+        Route::get('/export', 'OrderController@getExportOrder')->middleware('admin')->name('order.getExport');
+        Route::put('/export/{id}', 'OrderController@exportOrder')->middleware('admin')->name('order.export');
 
-        Route::get('/shipped', 'OrderController@getShippedOrder')->name('order.getShipped');
-        Route::put('/shipped/{id}', 'OrderController@shippedOrder')->name('order.shipped');
+        Route::get('/shipped', 'OrderController@getShippedOrder')->middleware('adminPage')->name('order.getShipped');
+        Route::put('/shipped/{id}', 'OrderController@shippedOrder')->middleware('adminPage')->name('order.shipped');
     });
 
 });;
